ntv2nub (rpclib) fixes, updates, work-arounds
- use uint64_t for POPU64
- Special NTV2Buffer::RPCDecodeClient() which does not call Allocate(), which would result in a new & incorrect memory location.
- For RPCDecode's with an NTV2Header for an NTV2Message, setting ConnectionID to 0 after decoding it.
On the server, after decoding, an NTV2Message call will be made.
On Linux, the first several I tested (NTV2GetRegisters & AUTOCIRCULATE_STATUS),
the driver would error if the connection id was present or non zero
(in the fOperatoin field of the NTV2_HEADER)
This is a work-around.  Should discover why the driver errors.

Issue #372

+++ ajantv2/includes/ntv2nubtypes.h
@@ -77,7 +77,7 @@ namespace ntv2nub
 
 	inline void POPU64 (uint64_t & outVal, const std::vector<uint8_t> & inArr, std::size_t & inOutNdx, const bool dontSwap = false)
 	{
-		uint32_t _u64(0);
+		uint64_t _u64(0);
 		UByte * _pU8(reinterpret_cast<UByte*>(&_u64));
 		_pU8[0] = inArr.at(inOutNdx++); _pU8[1] = inArr.at(inOutNdx++);
 		_pU8[2] = inArr.at(inOutNdx++); _pU8[3] = inArr.at(inOutNdx++);
